Replaced JWT with custom auth scheme based on HMACSHA256 session signing.
This follows crypto industry's recommendation of avoiding JWT because of the giant footgun that it is. The same thing can be done much more securely by just creating a secret key and signing the sessions.

I've also started tracking sessions in the database in this commit, which lets us invalidate any session by just deleting it from the database.